Refactor: divide ActionScreen's code(SelectSlotPage)
