+++ src/components/App.js
@@ -1,13 +1,18 @@
 import { SelectActionPage } from "../pages/SelectActionPage.js";
 import { OTPPage } from "../pages/OTPPage.js";
-import { SelectItemPage } from "../pages/SelectItemPage.js"
+import { SelectItemPage } from "../pages/SelectItemPage.js";
+import { SelectSlotPage } from "../pages/SelectSlotPage.js";
+
 import { ActionScreen } from "../pages/ActionScreen.js";
+
 import { createState } from "../core/core.js";
+import { performLockerAction } from "../services/api.js";
 
 const [currentPage, setCurrentPage] = createState("selectAction");
 const [selectedAction, setSelectedAction] = createState(null);
 const [userSession, setUserSession] = createState(null);
 const [selectedItem, setSelectedItem] = createState(null);
+const [selectedSlot, setSelectedSlot] = createState(null);
 
 const renderPage = () => {
   if (currentPage() === "selectAction") {
@@ -37,6 +42,28 @@ const renderPage = () => {
       setSelectedItem: setSelectedItem,
       onSelect: () => {
         setCurrentPage("selectSlot")
+      }
+    });
+  }
+
+  if (currentPage() === "selectSlot") {
+    return SelectSlotPage({
+      availableSlots: userSession().available_slots,
+      selectedSlot: selectedSlot,
+      setSelectedSlot: setSelectedSlot,
+      onSelect: async () => {
+        const res = await performLockerAction({
+          action: selectedAction(),
+          item: selectedItem(),
+          slot: selectedSlot(),
+        });
+        if (res.success) {
+          setCurrentPage("openLocker")
+
+        } else {
+          alert("사물함 동작에 실패했습니다. 다시 시도해주세요.")
+        }
+        
       }
     });
   }
