+++ src/pages/ActionScreen.js
@@ -6,48 +6,7 @@ const [phase, setPhase] = createState("select_item");
 const [selectedItem, setSelectedItem] = createState(null);
 const [selectedSlot, setSelectedSlot] = createState(null);
 
-export const ActionScreen = ({ action, userName, items, availableSlots }) => {
-    const renderSlotSelectionStoreOrReturn = () => {
-        const slotButtons = availableSlots.map((slot) => {
-            const { html, handlers } = Button({
-                id: `slot-${slot}`,
-                label: `${slot}번 칸`,
-                onClick: () => setSelectedSlot(slot)
-            });
-            return { html, handlers };
-        });
-
-        const { html: confirmHtml, handlers: confirmHandlers } = Button({
-            id: "confirm-perform",
-            label: "확인",
-            onClick: async () => {
-                console.log("실행 요청:", action, selectedItem(), selectedSlot());
-                const res = await performLockerAction({
-                    action: action,
-                    item: selectedItem(),
-                    slot: selectedSlot()
-                })
-
-                if (res.success) {
-                    setPhase("done");
-                } else {
-                    alert("사물함 동작에 실패했습니다. 다시 시도해주세요.")
-                }
-            }
-        });
-
-        return {
-            html: `
-                <div class="screen-container">
-                    <h2>사용할 칸을 선택하세요</h2>
-                    ${slotButtons.map(b  => b.html).join("")}
-                    <div class="action-button-wrapper">${confirmHtml}</div>
-                </div>
-            `,
-            handlers: Object.assign({}, ...slotButtons.map((b) => b.handlers), confirmHandlers)
-        };
-    };
-
+export const ActionScreen = ({ action, userName }) => {
     const renderSlotDisplayBorrowOrRetrieve = () => {
         const slot = selectedItem()?.slot;
         const slotInfo = slot ? `${slot}번 칸에서 꺼내주세요` : "다른 사물함에 있습니다.";
